Allow using the SDK for payment processing only
When the appropriate success and cancel redirects for Payment are added in the manifest
it is now possible to handle payment transactions without signing in with Connect ID beforehand.
The sign-in is handled by the payment backend instead.

This also gets rid of enabling Payment explicitly, it is alwaysn enabled when the correct
meta-data entries are present.

Change-Id: If3ae18f7ce11e049432a4df9ef00dcb5a987df06