Expose getAccessTokenFromCode function to support external browser
It should now be possible to use an external browser with public clients.
The developer has to register a schema for the redirect URI on an activity.
Then in that activity a call to getAccessTokenFromCode with the code
received in the Intent to open that Activity will call oauth/token and
store the access tokens in the client automatically.

Change-Id: I8ffcae82569f1ccbbee8a378af51ceaf990ba96d