Change version to 1.0.4
This version is only intended for MyContacts to test failing
deserialisation of tokens.
It will not have a public GitHub release and is not intended for the
general public as it will crash the SDK, something we avoid in the
official version.
Details of the discussion could be found in an email thread.