Fix the hotspot issue
The hotspot issue happens when the device is connected to a WiFi 4G
hotspot. In this case, the  Header Enrichment service will return the
MSISDN of the hotspot. To prevent that we won't to disable Header
Enrichment when the device is only using WiFi.

Since on Android there's not a reliable API to get if only the WiFi
adapter is used, we enable the Forced HE patch for all every time a
loadGif is performed. Then failing with an IOException during
Network.openConnection is a safe way to assume that the adapter is not
connected. In this case we return 404 for the loadGif and abort the HE flow.