Avoid requesting /oauth/revoke without tokens
A possible pattern could be that they call ConnectSdk.logout() without previously
being logged in, possibly to make a fresh start or something. It is worth mentioning
that this pattern does not producea crash in the app, just some annoying log entries
in the app and on the backend side.

Though this pattern looks like a misuse of the SDK, I think the SDK (and the backend
for that matter)  would benefit from not calling the revoke endpoint without tokens.
Otherwise, we might be snowed down with http 400 if many apps call
ConnectSdk.logout() unconditionally.