Change naming of payment link
After creating a new transaction, the webserver was passing a link to
the payment ui via a json field called `location`. Apparently, this has
caused 3rd party developers to misunderstand, and redirect their users
to the `Location` header returned from payment. This leads them to be
presented with the wrong payment UI.