Improve failure handling of CHANGE_NETWORK_STATE acquiring in Android 6.0
Android 6.0 has problems with auto-granting CHANGE_NETWORK_STATE
permission that is required for Forced HE and additional logic is
required to acquire it. This has been fixed in Android 6.0.1.
In case of an exception during requestNetwork() (CHANGE_NETWORK_STATE
was not granted) we should not crash, but just disable Forced HE.