Remove option to use fallback components directly (#192)
We expect SDK users to have a single entry point to the sdk
and provide additional fallback options in case if that entry
point fails. Before it was possible to use fallback options directly,
that was allowing the usage of some bad practises. Now this option
will be disabled.

This is a breaking change.